Create an event-controller & move logic of render event to it
